refactor(transaction): own foreign exchange domain
