+++ src/services/portfolio_transaction_processing_service/app/domain/transaction/fx/__init__.py
@@ -1,23 +1,23 @@
-"""Expose cross-capability FX and effective-processing compatibility contracts."""
+"""Expose canonical foreign-exchange transaction economics policies."""
 
-from .fx_baseline_processing import (
+from .baseline_processing import (
     UnsupportedFxRealizedPnlModeError,
     assert_fx_processed_event_valid,
     build_fx_baseline_processing_update,
     build_fx_processed_event,
 )
-from .fx_contract_instrument import (
+from .contract_instrument import (
     FX_CONTRACT_ASSET_CLASS,
     FX_CONTRACT_PRODUCT_TYPE,
     build_fx_contract_instrument_event,
     is_fx_contract_component_event,
 )
-from .fx_linkage import (
+from .linkage import (
     FX_DEFAULT_POLICY_ID,
     FX_DEFAULT_POLICY_VERSION,
     enrich_fx_transaction_metadata,
 )
-from .fx_models import (
+from .models import (
     FX_BUSINESS_TRANSACTION_TYPES,
     FX_CASH_LEG_ROLES,
     FX_COMPONENT_TYPES,
@@ -26,8 +26,8 @@ from .fx_models import (
     FX_SPOT_EXPOSURE_MODELS,
     FxCanonicalTransaction,
 )
-from .fx_reason_codes import FxValidationReasonCode
-from .fx_validation import FxValidationError, FxValidationIssue, validate_fx_transaction
+from .reason_codes import FxValidationReasonCode
+from .validation import FxValidationError, FxValidationIssue, validate_fx_transaction
 
 __all__ = [
     "FX_BUSINESS_TRANSACTION_TYPES",
