+++ src/services/portfolio_transaction_processing_service/app/domain/transaction/fx/baseline_processing.py
@@ -1,3 +1,5 @@
+"""Apply the governed baseline cost and realized-P&L policy to FX components."""
+
 from __future__ import annotations
 
 from decimal import Decimal
@@ -6,8 +8,9 @@ from portfolio_common.domain.transaction_control_codes import (
     normalize_transaction_control_code,
 )
 from portfolio_common.events import TransactionEvent
-from portfolio_common.transaction_domain.fx_models import FxCanonicalTransaction
-from portfolio_common.transaction_domain.fx_validation import validate_fx_transaction
+
+from .models import FxCanonicalTransaction
+from .validation import validate_fx_transaction
 
 FX_BASELINE_REALIZED_PNL_MODES = {"NONE", "UPSTREAM_PROVIDED"}
