+++ src/services/portfolio_transaction_processing_service/app/domain/transaction/fx/reason_codes.py
@@ -1,3 +1,5 @@
+"""Stable validation reason codes for foreign-exchange transactions."""
+
 from enum import StrEnum
