+++ tests/unit/services/query_service/test_test_manifest.py
@@ -53,7 +53,10 @@ def test_sell_contract_suite_includes_sell_query_contract_tests() -> None:
 
 def test_fx_contract_suite_includes_fx_contract_surfaces() -> None:
     fx_suite = get_suite("transaction-fx-contract")
-    assert "tests/unit/libs/portfolio_common/test_fx_validation.py" in fx_suite
+    assert (
+        "tests/unit/services/portfolio_transaction_processing_service/transaction/fx/"
+        "test_validation.py"
+    ) in fx_suite
     assert "tests/integration/services/query_service/test_transactions_router.py" in fx_suite
     assert (
         "tests/integration/services/persistence_service/repositories/test_repositories.py"
